SSAFY/django/관통프로젝트4/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.7
django-extensions==3.1.3
djangorestframework==3.12.4
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
pycodestyle==2.8.0
Pygments==2.10.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.1.0
wcwidth==0.2.5
